backend/pyproject.toml
[project]
name = "vezha-backend"
version = "1.0.0"
description = "VEZHA Digital API"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn[standard]>=0.27.1",
    "sqlalchemy[asyncio]>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.26.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.2.1",
    "black>=25.12.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.14",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
